feat(import_devices): support devices that don't use ids.id
Right now if a device has an ids field, that means that there are multiple devices that are fundimentally the same.
It expects an id field, that it will do replacement on the topic with.

This changes it so it doesn't require you to have the id, you could just set the topic or something.

Also chages from a try/catch to see if the ids field exists to if/else.